+++ source/objects/delayLine.hpp
@@ -15,7 +15,7 @@ public:
     // constructor and destructor
     DelayLine() = default;
 
-    DelayLine(float sampleRate, uint32_t size);
+    DelayLine(double sampleRate, uint32_t size);
 
     ~DelayLine();
 
@@ -45,7 +45,7 @@ public:
 
 private:
     // variables
-    float sampleRate = 44100;
+    double sampleRate = 44100;
     float* buffer = nullptr;
     uint32_t bufferSize;
     uint32_t readHead = 0;
